requirements.txt
djangorestframework==3.15.1
asgiref==3.9.1
dataclasses==0.6
Django==5.2.4
dnspython==2.7.0
pymongo==4.13.2
sqlparse==0.5.3
tzdata==2025.2
